The implementation before this change list keeps the ownership of memory that is used by peer connection instances in the peer connection manager. This means that if the peer connection manager is deleted before all peer connections it has created, these peer connections will be pointing to invalid memory.
The solution in this CL is to create a bundle of the memory that needs to be alive as long as there are any peer connections or peer connection manager instances. This bundle is scoped reference counted so that it is deleted only when there are no references to it. This enables the peer connection and manager to be deleted in any order.
Review URL: http://webrtc-codereview.appspot.com/246003

git-svn-id: http://webrtc.googlecode.com/svn@843 4adac7df-926f-26a2-2b94-8c16560cd09d