Optimized WebRtcIsacfix_NormLatticeFilterMa() function for iSAC fix for ARM Neon architecture with intrinsics and assembly code. The total iSAC codec speech improved about 3~5%.
Notes
(1) The Neon version after this optimization is not bit-exact with the generic
C version. The out quality, however, is not worse as verified by test vectors ouput,
and undertandably in theory (32bit x 32bit in Neon is more accurate than the approximation
C code in the generic version).
(2) In Android, a isac neon library will be built. Along with some new function structures,
it is partly for preparation of introducing a run time detection of Neon architecture soon.
Review URL: http://webrtc-codereview.appspot.com/268016

git-svn-id: http://webrtc.googlecode.com/svn@1192 4adac7df-926f-26a2-2b94-8c16560cd09d