Rolled back YASM to versions we know are working.
The idea here is to attempt to provide a short-term solution to the flaky YASM failures on the windows bots. We could run with this a bit, and if the errors cease we know it is a bug in YASM and can perhaps proceed to report a bug on YASM. This is probably not a good long-term solution since I guess we want to use the same tools as Chrome.

This effectively rolls back the YASM version changes of r1755 (see http://code.google.com/p/webrtc/source/diff?spec=svn1755&r=1755&format=side&path=/trunk/DEPS)

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/453005

git-svn-id: http://webrtc.googlecode.com/svn@1886 4adac7df-926f-26a2-2b94-8c16560cd09d