Fix how we were using TbInterfaces and disallow operator=() and the copy constructor. The reason is that this will cause a crash:
TbInterfaces foo = TbInterfaces("blah");


It relies on the generated copy constructor (or assignment operator), which copies the
pointer values from a temporary object.  After |foo| in this case has been initialized
with values from the temporary object, the temp goes out of scope and is deleted.
The result is that |foo| has been initialized with pointers do a deleted object.

Also fixing expectations for the return value of VoE Release() calls after I checked
in my change that makes the VoiceEngine per-object ref counted and not per-interface.
Review URL: https://webrtc-codereview.appspot.com/509005

git-svn-id: http://webrtc.googlecode.com/svn@2128 4adac7df-926f-26a2-2b94-8c16560cd09d