Fix a potential Chrome crash in audioproc.
There are Chrome crashes which appear to be occurring during some kind
of teardown. We might be able to avoid them by locking in the destructor.

On the other hand, this might have no impact, but at least isn't a bad
thing to do.

BUG=chromium:145341

Review URL: https://webrtc-codereview.appspot.com/768005

git-svn-id: http://webrtc.googlecode.com/svn@2687 4adac7df-926f-26a2-2b94-8c16560cd09d