Rewrote menu handling for vie custom call.
The intended trajectory of this patch is to abstract out all i/o for custom_call.
The reason is that kjellander@ will need to be able to configure custom calls using
flags, and using the same framework to gather all input gathering to a single place
will make this a lot easier.

This patch focuses on choices. The next will focus on field entries, like "enter
ip address" or "enter port number."

BUG=
TEST=Manually tested all menus in custom call, ran new unit tests.

Review URL: https://webrtc-codereview.appspot.com/757005

git-svn-id: http://webrtc.googlecode.com/svn@2758 4adac7df-926f-26a2-2b94-8c16560cd09d