Adding video_coding_integrationtests test.
These changes makes it possible to run this tool with some gtest additions in an automated manner on the buildbots.

This test was previously known as video_coding_test, which is an
integration test that is mostly used as a development tool.

Parts of this test should be extracted and kept as a separate
development tool, but that's something for a future CL.

I also refactored the old command line parsing to use gflags instead.

Previous code from the following tests were merged into
video_coding_integrationtests and video_coding_unittests:
* video_codecs_test_framework_integrationtests
* video_codecs_test_framework_unittests
So these targets are now gone.

BUG=none
TEST=trybots passing + Executing video_coding_integrationtests on Linux, Mac and Windows since it's not currently added to the trybots. I ran with a couple of different combinations of settings.

Review URL: https://webrtc-codereview.appspot.com/933026

git-svn-id: http://webrtc.googlecode.com/svn@3176 4adac7df-926f-26a2-2b94-8c16560cd09d