Add <(DEPTH) to global includes
With http://review.webrtc.org/1064004 we got use of headers in testing/ for the first time in production code (which is what gtest_prod.h is meant for). This showed that when WebRTC is built inside Chrome, the include path doesn't include the top-level directory, so testing/gtest/ could not be found.

By adding <(DEPTH) to the WebRTC global include path list in common.gypi, this is resolved.

Having this directory in the global include path list will also make it possible for us to use full paths for common third party libraries, which should be something we aim for.

BUG=none
TEST=Successfully compiled the  webrtc_test_tools target on Linux in a Chromium checkout with third_party/webrtc replaced by ToT trunk with this patch applied (with Python 2.6 installed, which is needed to get the pyautolib target generated).

Review URL: https://webrtc-codereview.appspot.com/1082004

git-svn-id: http://webrtc.googlecode.com/svn@3421 4adac7df-926f-26a2-2b94-8c16560cd09d