Don't upsample the capture signal early.
* Remove the unneeded _mixingFrequency.
* Rename CheckForSendCodecChanges to better elucidate its function.
* Remove an unnecessary memcpy.

Upsampling should be done late in the chain. This is practically relevant
on mobile, where the capture rate is fixed at 16 kHz. When using Opus, the
signal was upsampled to 32 kHz and was no longer compatible with AECM, which only supports up to 16 kHz.

NEEDS_QA=true
TEST=run calls with a variety of capture device rates and codecs
BUG=chromium:178040,webrtc:1446

Review URL: https://webrtc-codereview.appspot.com/1146004

git-svn-id: http://webrtc.googlecode.com/svn@3594 4adac7df-926f-26a2-2b94-8c16560cd09d