Add GYP target for WebRTC Video demo for Android.
Add a build target for the Video demo app for Android that only
exists when OS=='android' during build.

Note that this doesn't solve webrtc:1029, it's more like a workaround
waiting for the complete solution, which is to great a proper GYP target
that doesn't involve an action and an external script.

BUG=1029
TEST=Built successfully with:
source build/android/envsetup.sh
gclient runhooks
ninja -C out/Debug
Also verified the target is not present when OS is not 'android'.

Review URL: https://webrtc-codereview.appspot.com/1286004

git-svn-id: http://webrtc.googlecode.com/svn@3769 4adac7df-926f-26a2-2b94-8c16560cd09d