Update iOS build script to run on bots.
The script previously only executed build/gyp_chromium to
generate projects. This is not sufficient since the
supplement.gypi has not yet been created (happens in
gclient runhooks). Therefore this CL changes it to run
gclient runhooks instead (marginally slower).
Otherwise the bots would have to generate the projects
twice, which is inefficient.

Annotations are also added to make the upcoming buildbot
steps more readable when the script is running as an
annotated step.

TEST=local compilation on OS X.
BUG=none

Review URL: https://webrtc-codereview.appspot.com/1342005

git-svn-id: http://webrtc.googlecode.com/svn@3878 4adac7df-926f-26a2-2b94-8c16560cd09d