Drop Virtual webcam check script as moved into buildbot scripts.
Having this script being located in the WebRTC repo doesn't make sense
since it has no connection to the source code.
Updating this script should apply to all build configurations and since
this script will be used for Chromium builders, we'll end up with having
to wait for a new WebRTC revision to be rolled in DEPS before it's updated.

TEST=none
BUG=none
TBR=phoglund

Review URL: https://webrtc-codereview.appspot.com/1444006

git-svn-id: http://webrtc.googlecode.com/svn@3992 4adac7df-926f-26a2-2b94-8c16560cd09d