Add precompiled Chromium's Android tools needed for testing.
Since these depend on Chromium base, we cannot compile them
from source in WebRTC. We'll keep a precompiled copy to be
able to run Android tests.

These binaries are based on Chromium r206006 and are
compiled in Release mode on Ubuntu Precise.

TEST=Tested in a WebRTC workspace to execute a test APK
on an Android device.
BUG=1882
R=henrike@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1670004

git-svn-id: http://webrtc.googlecode.com/svn@4235 4adac7df-926f-26a2-2b94-8c16560cd09d