Fixed the AGC and interface problems on the new path.
In order to make the AGC work properly, we need to cache the volume value passed by the callback, compare it with the value returned by shared->transmit_mixer()->CaptureLevel(). If they are the same, we need to return 0 to indicate no volume needs changing, otherwise return the new volume.
By doing this, we avoid setting the volume all the same, which allows the users to change the volume manually.

This patch also fixes some minor issues with the interfaces too: make the int channel[] const, and correct the order of the input params in channel::Demultiplex.

BUG=[2134]
TEST=compile && manual AGC test
R=andrew@webrtc.org, tommi@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1839004

git-svn-id: http://webrtc.googlecode.com/svn@4394 4adac7df-926f-26a2-2b94-8c16560cd09d