Improved error messages when binaries are missing. Also stderr = stdout now.
Now that these scripts are called from browser tests, we need to print everything on stdout since the tests will throw away stderr when invoking programs. I chose to assign sys.stderr to sys.stdout. Otherwise I would have missed stuff like parser.error, which print to stderr.

The error message will get improved because the old code did not catch the case when the binary was missing, which lead to a very confusing error when that was the case. This gets fixed now.

BUG=
R=kjellander@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1886004

git-svn-id: http://webrtc.googlecode.com/svn@4416 4adac7df-926f-26a2-2b94-8c16560cd09d