Rename EngineTest to CallTest.
There's no real notion of VideoEngine left in these classes. They're
end-to-end tests built on Call, so CallTest makes more sense.

This also contains a modification to RtpRtcpObserver moving the
responsibility of creating the event that signals when the observation
is complete to RtpRtcpObserver. New tests are about to be introduced and
this will reduce code duplication.

BUG=
R=stefan@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2258005

git-svn-id: http://webrtc.googlecode.com/svn@4793 4adac7df-926f-26a2-2b94-8c16560cd09d