Rewriting the SoundMeter class to be RMS and be encapsulated differently
This CL changes the SoundMeter to be root-mean-square.
It also changes the interface between the meter and the display to be based on the display calling down to the meter rather than the meter calling up to the display.

A graphic display of the results is also added.

BUG=
R=cwilso@google.com, dutton@google.com, henrika@webrtc.org, juberti@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/5439004

git-svn-id: http://webrtc.googlecode.com/svn@5256 4adac7df-926f-26a2-2b94-8c16560cd09d