Output logs to stderr from voe_cmd_test by default.
Add a flag --log_file which produces the existing behaviour of dumping
logs of all severities to a file. By default, warnings and errors will
now be output to stderr. This is generally more useful for the testing
done with voe_cmd_test.

TESTED=logs output to stderr by default and to the usual file when the
flag is specified.

R=tnakamura@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/6849005

git-svn-id: http://webrtc.googlecode.com/svn@5409 4adac7df-926f-26a2-2b94-8c16560cd09d