This is to roll a more recent Chromium, which contains latest Clang, so as to be able to roll Opus 1.1, which will the next step.
There are uninitializion problem with normal_asyn_test.cc. This is fairly easy to solve and therefore is included in this CL.

The following is a memo on the selection of the version to roll. It may be a reference for similar missions.

How was this version picked?

1. The whole purpose of this work is to update to Clang to be able to compile Opus 1.1. In Chromium, Clang got updated to 198389 at r244540.

2. From r245412, gyp_chromium requires "tools\find_depot_tools.py". However, WebRTC does not sync up the root of folder "tools". An issue has been created to Chromium on this.

... So the version must be a good version between r244540 and r245411 (inclusive)

BUG=

TEST=passes all trybots
R=kjellander@webrtc.org, stefan@webrtc.org, tina.legrand@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/7569005

git-svn-id: http://webrtc.googlecode.com/svn@5436 4adac7df-926f-26a2-2b94-8c16560cd09d