Echo cancellation functions docs: Follow style guide w.r.t. placement of *
The style guide says to use "void* x", not void *x", and the code in
these files already do so, but the comments do not. Fix that.

Also, in the interest of reducing eye strain, I fixed the vertical
alignment in a small number of cases.

BUG=
R=aluebs@webrtc.org, andrew@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/12509004

git-svn-id: http://webrtc.googlecode.com/svn@6101 4adac7df-926f-26a2-2b94-8c16560cd09d