Wire up --force_fieldtrials for vie_auto_test and for test targets linking with test/test.gyp:{test_main|test_support_main}
This allows use of webrtc field trials  and opens up the possibility to try the different code paths when running the unit tests by wiring them up to a --force_fieldtrials.

Tested: running a test target that links with the above with a flag --force_fieldtrials=invalid leads the test to crash.

BUG=crbug/367114
R=mflodman@webrtc.org, phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/19439004

git-svn-id: http://webrtc.googlecode.com/svn@6181 4adac7df-926f-26a2-2b94-8c16560cd09d