audio_processing: Forces extended filter to be used in splitting filter test.
The behavior differ between "normal" and "extended" modes when using AEC. In the extended filter mode nothing is processed until we have received a farend frame. This is exactly what is needed in this part of the splitting filter test.
On Android, we do not use the normal mode, which made the test to fail.

BUG=3445
R=kwiberg@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/12679004

git-svn-id: http://webrtc.googlecode.com/svn@6368 4adac7df-926f-26a2-2b94-8c16560cd09d