Fixes a bug causing NACKs to be dropped excessively at the send-side.
This was introduced in r6472 where the target bitrate was changed to be stored in bits/s instead of kbits/s, but the storage type was accidentally left as uint16_t. This caused the bitrate to be truncated, which at times causes NACKs to be dropped due to insufficient bitrate available.

BUG=3518
TEST=Tested in Chrome, trybots and verified that it fixes the bug in vie_auto_test loopback test.
R=mflodman@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/21739004

git-svn-id: http://webrtc.googlecode.com/svn@6544 4adac7df-926f-26a2-2b94-8c16560cd09d