Fixes a potential BWE clock mismatch bug.
Since libjingle provides a packet arrival timestamp to webrtc, and the clock in remote bitrate estimator and the clock used for packet arrival timestamp can be different. This can cause the bandwidth estimator to malfunction.

This CL changes the remote bitrate estimator so that packet arrival timestamps never are compared to the time taken from the internal clock.

BUG=3527
R=mflodman@webrtc.org, pbos@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/20789004

git-svn-id: http://webrtc.googlecode.com/svn@6571 4adac7df-926f-26a2-2b94-8c16560cd09d