GN: Add PRESUBMIT.py check for GN changes + default bots.
Add the GN trybots to the default set and also set them
to be the only bots to run if a CL contains only BUILD.gn
changes.

Update Python exclusions in general and fix a few of the lint
warnings.
The ones in python_charts needs to be disabled since those variables
are actually used when passed via vars() to the template.

BUG=None
TEST=git cl presubmit with the following cases:
A CL with two .gyp changes.
A CL with no changes in .gyp* files.

R=niklas.enbom@webrtc.org, phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/18719004

git-svn-id: http://webrtc.googlecode.com/svn@6834 4adac7df-926f-26a2-2b94-8c16560cd09d