modules/audio_processing: Moves declaration of kDelayDiffOffsetSamples
audio_processing did not compile when aec_untrusted_delay_for_testing=1 was set. The constant kDelayDiffOffsetSamples was declared only for Mac when WEBRTC_UNTRUSTED_DELAY was automatically turned on.

Moving the declaration outside the ifdef makes it build with the flag on for any platform.

BUG=3673
TESTED=locally and trybots
R=andrew@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/22049004

git-svn-id: http://webrtc.googlecode.com/svn@6866 4adac7df-926f-26a2-2b94-8c16560cd09d