Pass --verbose to gclient sync of Chromium
In r6938 the switch to using Chromium's Git repo was
deployed. However this fails on the bots since their timeout
for steps without output is 1200 seconds, which is not enough
to checkout the large Chromium Git repo.
Adding --verbose will print more output, thus getting a longer
timeout that should be enough for the runhooks step to complete.

BUG=2863, chromium:339647
TEST=None
TBR=phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/15209004

git-svn-id: http://webrtc.googlecode.com/svn@6939 4adac7df-926f-26a2-2b94-8c16560cd09d