Increase verbosity for gclient sync of Chromium
In r6939 the --verbose flag was passed to the problematic
(approx 2.2GB large) gclient sync of Chromium's src.git repo.
However the bots are still hitting killed sync jobs due to
lack of output. This is a speculative attempt to provoke
even more logging, in order to trigger buffer flushing for
the buildbot execution.

BUG=2863, chromium:339647
TEST=Ran gclient runhooks locally with CHROME_HEADLESS=1 set.
TBR=phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/21269004

git-svn-id: http://webrtc.googlecode.com/svn@6940 4adac7df-926f-26a2-2b94-8c16560cd09d