Fix the RTC+Chromium GN build.
LOGGING_INSIDE_WEBRTC was being set in the inherited config, whereas in the GYP build this define is not inherited. This caused duplicate logging macros to be defined in Chrome files dependening on WebRTC targets.

Move LOGGING_INSIDE_WEBRTC to the common config (non-inherited).

TBR=kjellander@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/25449004

git-svn-id: http://webrtc.googlecode.com/svn@7122 4adac7df-926f-26a2-2b94-8c16560cd09d