Remove root_dir variable from DEPS + enforce rename.
Update DEPS to no longer have the root_dir variable.
Add a script that detects if the user have a solution named
'trunk' and explains what needs to be done to change it to 'src'.

The reason for this change is that bot_update doesn't support
custom_vars in solutions and Chrome infra is trying to get
rid of them entirely in the future.

The bots are already using a solution named 'src' so they
won't run into this error during runhooks.

BUG=3534
TESTED=Ran the script with a .gclient containing a solution
named 'trunk' and one named 'src'. Also tested the presence
of the custom_vars dict and not.

R=andrew@webrtc.org, hinoka@chromium.org

Review URL: https://webrtc-codereview.appspot.com/30619004

git-svn-id: http://webrtc.googlecode.com/svn@7405 4adac7df-926f-26a2-2b94-8c16560cd09d