PRESUBMIT: Only notify GN changes for GYP files in webrtc/*
We don't maintain a BUILD.gn file for talk/ since it's a part
of Chromium:
https://code.google.com/p/chromium/codesearch#chromium/src/third_party/libjingle/BUILD.gn
Because of this, it's confusing to get warnings about updating
a GYP file in talk/ from the PRESUBMIT check.

TESTED=Successsfully ran git cl presubmit with this change
applied on top of a CL containing changes in .gyp files.

R=niklas.enbom@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/27189004

git-svn-id: http://webrtc.googlecode.com/svn@7737 4adac7df-926f-26a2-2b94-8c16560cd09d